chore: checkpoint farm detail plan polish
